[native] Create a place where logs are stored and can be displayed
Summary:
The logs are stored in a new context. We could use Redux for that, but the benefits would be quite limited because we don't need to persist the logs. We can always consider changing the approach and persist them, but they are a simple replacement to the alerts, which aren't persisted as well.

The logs are displayed on a new staff-only screen accessible from the profile screen. It is possible to clear them and copy to a clipboard.

Test Plan: Added a button that results in adding ew logs. Tested if they are displayed correctly on the screen, can be copied, and cleared.

Reviewers: kamil, angelika

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14295