[native_rust_library] avoid triggering backup upload before cleanup
Summary: I discovered this when testing this for wallet users and looking into logs. We might want to start uploading in the middle of the cleanup process, with this easy change we'll wait for deleting already uploaded files first.

Test Plan: Check logs that there is now upload until files are deleted.

Reviewers: bartek, tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14294