[native] Display a badge on the profile tab when there are some logs
Summary:
We should avoid a case where the logs are being ignored, but we should also make them visible. In order to do that we can use a badge that shows the number of logs only for stall users.

https://linear.app/comm/issue/ENG-10140/replace-staff-alerts-with-a-log

{F4045274}

{F4045275}

Depends on D14295

Test Plan: Checked if the number of logs is displayed and clearing the logs causes the badge to disappear.

Reviewers: kamil, angelika, ashoat

Reviewed By: kamil, ashoat

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14296