[native] Cancel animations when gesture is started in FullScreenViewModal
Summary:
https://linear.app/comm/issue/ENG-8150/convert-fullscreenviewmodal-to-reanimated-2-syntax

As mentioned here: https://phab.comm.dev/D14305 in the original implementation there are a couple of places that stop animation if gesture is active. In this diff I added this logic.

The builds will fail due to some leftover reanimated v1 code. I'll remove it in the next diffs for diffs readability.

Depends on D14308

Test Plan: Test after migration: try to start a gesture when animation is in progress

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14309