[keyserver] Update unread checks to use new last_message_for_unread_check column
Summary: This is a no-op, as the column currently has the same value as `last_message`. It will be changed in later diffs.

Depends on D14329

Test Plan: Tested in combination with later diffs, to confirm that it's possible for some message types to not set a thread to unread

Reviewers: tomek, angelika

Subscribers: