[keyserver] Introduce latestMessageForUnreadCheck to LatestMessageInfo
Summary: This diff is a no-op; it simply moves the setting of this value "upstream" a bit. Later diffs will change the actual value.

Depends on D14332

Test Plan: Tested in combination with later diffs, to confirm that it's possible for some message types to not set a thread to unread

Reviewers: tomek, angelika

Subscribers: