[lib] Fix DM push notifs code to fetch message from SQLite instead of Redux
Summary:
This code fetches a message that is reacted to in order to check its author, but it's currently set up to only check Redux.

Instead, we should query SQLite in case the message is not currently in Redux. `getLatestMessageEdit` also contains a fallback to check Redux if the data is not in SQLite yet.

Technically `getLatestMessageEdit` doesn't return just the `RawMessageInfo` that was queried for; rather, it returns the latest edit (only applies to `messageTypes.TEXT`). I think this is fine in this case, since we only need to use the `creatorID`. It might be preferable in a future use case, but hard to know for sure.

Depends on D14337

Test Plan: Tested in combination with later diffs, to confirm that it's possible for some message types to not set a thread to unread

Reviewers: tomek, angelika

Differential Revision: https://phab.comm.dev/D14339