[lib] Add messagesNotifyTypes to ProcessDMOpsPayload
Summary: We want to get access to this data in reducers. Since it requires an `await`, we need to determine it prior to the dispatch.

Depends on D14341

Test Plan: Tested in combination with later diffs, to confirm that it's possible for some message types to not set a thread to unread

Reviewers: tomek, angelika

Subscribers: