[native] Navigate directly to the secondary device connection screen
Summary:
Currently, when a user scans the QR code using a camera app or clicks a deep link, we show a screen where a QR can be scanned or a URL can be pasted. This can be improved by using the value that was just received.

https://linear.app/comm/issue/ENG-10168/improve-qr-login-flow-on-dev

{F4037513}

Depends on D14315

Test Plan: Click a deep link to the QR login and check that a connection screen is displayed instead of the input field.

Reviewers: kamil, bartek, ashoat

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14326