[web][push-notif] free memory in `encryptNotificationWithOlmSession` and `decryptWithSession`
Summary:
Fixing issue described in [ENG-10144#comment-16a38c06](https://linear.app/comm/issue/ENG-10144/investigate-wasm-memory-leaks#comment-16a38c06).

Tracked in [ENG-10182](https://linear.app/comm/issue/ENG-10182/fix-new-olmaccount-usage-on-service-worker) and [ENG-10184](https://linear.app/comm/issue/ENG-10184/fix-new-olmsession-usage-on-service-worker)

Make sure all objects created in this function are freed. This is not as easy as in keyserver to create a generic function that handles it, as logic on the client is a lot more complicated and different between content and notif.

Depends on D14348

Test Plan: Tested keyserver and DM notifs.

Reviewers: tomek, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14349