[web][push-notif] free memory after calling `getNotifsCryptoAccount`
Summary:
Fixing issue described in [ENG-10144#comment-16a38c06](https://linear.app/comm/issue/ENG-10144/investigate-wasm-memory-leaks#comment-16a38c06).

Tracked in [ENG-10182](https://linear.app/comm/issue/ENG-10182/fix-new-olmaccount-usage-on-service-worker).

Make sure all objects created in this function are freed. This is not as easy as in keyserver to create a generic function that handles it, as logic on the client is a lot more complicated and different between content and notif.

I tried to avoid returning `olm.Account` from this function, but it was not something we can easily do, I tried multiple solutions and all of them was a lot longer and less readable. Additionally, This really complex logic that was stabilizied for a long time and I don't want to tinker with this code.

Depends on D14349

Test Plan: Try using app for a while

Reviewers: tomek, bartek