[web] Extract the common logic from the QR auth
Summary:
The same logic is used in two places, which means we can extract it. We can't simply share the logic with native, where a `focusEffect` is used, but this isn't a big issue because native isn't affected by the bug we're fixing.

https://linear.app/comm/issue/ENG-10200/secondary-logout-in-signed-device-list-is-not-working-on-web

Test Plan: Test if logging in using the QR code works.

Reviewers: kamil, bartek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14375