[web] Create an unauthenticated Tunnelbroker connection only when socket connection becomes unauthorized
Summary:
Wait with starting the QR flow for the socket to become unauthorized. This will happen when `userID` becomes `null` during the logout.

This diff delivers a little broken experience, because the QR can be undefined for some time before the correct value is set - this will be addressed as a part of rebasing over D14368.

https://linear.app/comm/issue/ENG-10200/secondary-logout-in-signed-device-list-is-not-working-on-web

Depends on D14375

Test Plan: Logged in and out using the QR flow. Made sure that the logout procedure was successful and the new QR code could be used to log in.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14376