[native][web] Retry QR code generation
Summary:
When QR generation fails, we should retry it. We're doing it with some timeout to avoid making the calls too frequent.

https://linear.app/comm/issue/ENG-10209/seeing-error-failed-to-generate-qr-code-message-task-cancelled

This diff also introduces a loading state for QR on native:
{F4127207}

Test Plan:
Modified the QR generation call to success infrequently and checked if it was retried.

Tested the following scenario both on native and web:
1. Open the QR screen on a logged out device
2. Log in
3. Log out
4. Log in again
Before this diff, this fails. After this diff, a loading state of the QR code is shown, and then the correct QR code is displayed.

Reviewers: kamil, bartek

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14383