[lib] update Tunnelbroker's `sendMessage` to be based on ref
Summary:
Part of [ENG-10186](https://linear.app/comm/issue/ENG-10186/primary-device-unable-to-add-secondary-device-adding-device-failed#comment-f452eecd) and overall improvement.

There are a couple of places in the codebase where we try to send a message and it could fail with `Tunnelbroker not connected` but only because using an outdated or not yet re-rendered version of `sendMessage`, for example:
1. There is an asynchronous `useCallback` which is doing a lot of things, and after calls `sendMessage`, it starts its execution when Tunnelbroker is disconnected.
2. During execution, a socket connection is established.
3. In the callback mentioned above, `sendMessage` is called but it fails because it is bound to the old `socketState` even though the socket is connected. This diff fixes this.

Not sure if this isn't too hacky, curious about alternatives. We still want to keep the state to return it and allow adding it to dependency where needed.

Test Plan: Simulate the case descrbed above

Reviewers: tomek, bartek, angelika

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14384