[native] Remove tipHorizontalOffset from NUXTipsOverlay
Summary:
`tipHorizontalOffsetRef` has a reanimated v1 value, but from reading the code I can deduce that this value is set and
not read anywhere. Therefore it can be removed.

Also `onTipContainerLayout` just sets `tipHorizontalOffsetRef`  so it can also be removed.

Depends on D14410

Test Plan: flow, eslint, login to a new account and see that the tips at the beginning are working

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14411