[native] Fall back to v1 login after `use_v1_flow` response
Summary:
When a user has a backup but their device list doesn't have a signature, try to restore an account, and if that fails, fall back to the v1 login.

https://linear.app/comm/issue/ENG-10220/restore-flow-consider-adding-2nd-fallback-to-v1

Test Plan: Tested for password and SIWE users. Create an account, delete a signature from the most recent device list. Try to restore - this fails without this diff, but after this diff it succeeds because the v1 fallback is used.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14412