[native] Avoid running the handler multiple times
Summary:
The effect contains some async logic that is run based on a condition that is later modified in the effect. This means that it is possible to run this effect multiple times before the condition is noticed to be true (without this change the logic runs usually 4 times).

https://linear.app/comm/issue/ENG-10271/generating-backups-for-siwe-users-who-used-v1-fallback-doesnt-work

Test Plan: Check if the logic is run just once, by adding some console logs. This code can be run when a SIWE user doesn't have a backup.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14416