[native] Remove background identity login handler
Summary:
When this handler is present, clearing the state for SIWE users results in forcing a logout, which interferes with sensitive data cleanup - that sometimes results in the data staying in the store.
We probably no longer need this handler because we can assume that all the users are logged in to identity. I might be wrong, though, so this diff is more like a proposal and a place for discussion.

https://linear.app/comm/issue/ENG-10271/generating-backups-for-siwe-users-who-used-v1-fallback-doesnt-work

Depends on D14416

Test Plan:
Wipe state and kill the app. Verify that the alert store doesn't contain recent alert info.

Reviewers: kamil, bartek