[native] navigate to nux tip on modal close if user opened CommunityJoinerModal from bottom sheet
Summary:
If the user opened the modal by accepting the bottom sheet prompt, we'll still want to show them the nux tip when they close the modal so that they can find it in the future.

Depends on D14420

Test Plan:
when i open the modal from the community drawer and then close it, nothing happens.

but if i open the modal from the bottom sheet and then close it, i see the NUX tip indicating where to find the directory

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D14421