[native] Schedule backup generation and upload after saving SIWE secrets
Summary:
When a SIWE user logs in using a v1 fallback, they don't have SIWE secrets saved in the DB. They are set after an interaction with a modal, but that doesn't trigger backup generation and upload. These operations were performed only after the app was restarted - when the app starts, we would read the data saved in the previous session.

This diff solves this issue by exposing a function that can be used to trigger backup generation and upload and using this function after user provides the missing data.

https://linear.app/comm/issue/ENG-10271/generating-backups-for-siwe-users-who-used-v1-fallback-doesnt-work

Depends on D14417

Test Plan: Log in using v1 fallback. Sign a backup and check on AWS if the backup appears. Wipe state and kill the app and then restore - verify that the v2 is used by noticing that backup decrypt step is performed.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14425