[lib] launch connect farcaster prompt
Summary: This work has been complete for 9 months, but looks like it was never launched. Flipped `DISABLE_CONNECT_FARCASTER_ALERT` to false and removed the `isDev` check in `shouldSkipConnectFarcasterAlert`. Doesn't seem that annoying to see the bottom sheet when devs log in on mobile, but happy to add the `isDev` check back if we think that's better.

Test Plan:
confirmed that the bottom sheet appears on login now

NOTE: didn't test the actual connect farcaster flow. i assume this has been tested already

Reviewers: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D14444