[web][keyserver] log Olm memory usage changes
Summary:
Some context in [ENG-10393](https://linear.app/comm/issue/ENG-10393/log-wasm-memory-usage-on-keyserver#comment-b874b5ec).

Those metrics are helpful for two reasons:
1. `get_total_memory` - gives us insight into total memory allocated, spikes in this value means that we might have a lot of promises running at once which allocated a lot of resources - I described this theory in [ENG-10393#comment-b874b5ec](https://linear.app/comm/issue/ENG-10393/log-wasm-memory-usage-on-keyserver#comment-b874b5ec).
2. `get_used_memory` - if this value grows over time we're experiencing memory leaks.

Looking at `transform-remove-console` config we exclude `error` so should be fine.

For the web, we might expect used memory to grow when using app because we're keeping sessions in memory, which means after starting conversation with new user used memory should grow, but only about the size of Session object. The logs could help us discover wheter in-memory store isn't to big, and if so - refactor this to avoid keeping all sessions in-memory.

Depends on D14484

Test Plan:
1. Tested cronjob to print some results when needed (after commenting `free`).
2. Tested web to print results.

Reviewers: bartek, tomek, ashoat

Differential Revision: https://phab.comm.dev/D14485