[web][keyserver] log Olm memory usage changes
Summary:
Some context in [ENG-10393](https://linear.app/comm/issue/ENG-10393/log-wasm-memory-usage-on-keyserver#comment-b874b5ec).

Those metrics are helpful for two reasons:
1. `get_total_memory` - gives us insight into total memory allocated, spikes in this value mean that we might have a lot of promises running at once which allocated a lot of resources - I described this theory in [ENG-10393#comment-b874b5ec](https://linear.app/comm/issue/ENG-10393/log-wasm-memory-usage-on-keyserver#comment-b874b5ec).
2. `get_used_memory` - if this value grows over time we're experiencing memory leaks (except for session creation on the web, see below).

For the web, we might expect used memory to grow when using the app because we're keeping sessions in memory, which means after starting a conversation with a new user used memory should grow, but only for the size of the Session object. The logs could help us discover whether the in-memory store isn't too big, and if so - refactor this to avoid keeping all sessions live at once.

Looking at the `transform-remove-console` config we exclude `error` so should be fine.

Depends on D14484

Test Plan:
1. Tested cronjob to print some results when needed (after commenting `free`).
2. Tested web to print results.

Reviewers: bartek, tomek, ashoat

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14485