[ci] Update mac runner image in github workflows
Summary:
https://linear.app/comm/issue/ENG-10412/apple-sdk-upcoming-minimum-requirement

Apps uploaded to App Store Connect must be built with Xcode 16 or later using an SDK for iOS 18.
So we need to upgrade our CIs to use Xcode 16.
macos-15 runner uses Xcode 16 by default, so upgrading it should solve the problem.

Test Plan:
Build the app on my local machine with Xcode 16 (I had it upgraded long ago).
Then land and verify that all github actions passed?
If we want to be sure that it works before landing  we can try making a PR to gh and run actions on the PR?

Reviewers:
tomek, bartek

Subscribers: