[lib] Fix bug in ChatThreadItemLoaderCache for simultaneous loadMostRecent calls
Summary: If `loadMostRecent` is called multiple times, then `this.loadingState` will track the most recent call. We check `this.loadingState` here to make sure an earlier call does not set `this.currentState`, but we should `break` regardless to make sure that the earlier call returns once all of its items are resolved.

Test Plan: I tested this along with the rest of the stack by making sure that everything still worked on my physical device with a release build, and that performance was not degraded

Reviewers: tomek, angelika

Differential Revision: https://phab.comm.dev/D14493