[lib] Only use threadInfo.creationTime for lastUpdatedTime if no messages
Summary:
Currently, we fall back to `threadInfo.creationTime` in cases where none of the messages return a concrete time.

I think that behavior isn't great. It tends to happen when we only have a single most recent message for a given thin thread, but in those cases `threadInfo.creationTime` can be way earlier than the recent messages.

This diff changes the behavior so that if none of the messages return a concrete time, we'll fall back to the most recent message timestamp instead. We'll only use `threadInfo.creationTime` if we have no alternatives.

Depends on D14493

Test Plan: I tested this along with the rest of the stack by making sure that everything still worked on my physical device with a release build, and that performance was not degraded

Reviewers: tomek, angelika

Differential Revision: https://phab.comm.dev/D14494