[native] Fix "hermes contains bitcode" error
After @graszka22 updated us to use Xcode 16 in D14487, the build [failed to upload](https://github.com/CommE2E/comm/actions/runs/14121530404/job/39562551833) to App Store Connect due to this error:

```
=============
UPLOAD FAILED with 1 error.
=============
2025-03-28 04:20:10.680 *** Error: Error uploading 'output/Comm.ipa'.
2025-03-28 04:20:10.680 *** Error: Validation failed Invalid Executable. The executable 'Comm.app/Frameworks/hermes.framework/hermes' contains bitcode. (ID: d18ef6a2-5468-40e8-b7d7-e6be3ef9e245) (409)
 {
    NSLocalizedDescription = "Validation failed";
    NSLocalizedFailureReason = "Invalid Executable. The executable 'Comm.app/Frameworks/hermes.framework/hermes' contains bitcode. (ID: d18ef6a2-5468-40e8-b7d7-e6be3ef9e245)";
    NSUnderlyingError = "Error Domain=IrisAPI Code=-19241 \"Validation failed\" UserInfo={status=409, detail=Invalid Executable. The executable 'Comm.app/Frameworks/hermes.framework/hermes' contains bitcode., id=d18ef6a2-5468-40e8-b7d7-e6be3ef9e245, code=STATE_ERROR.VALIDATION_ERROR, title=Validation failed, NSLocalizedFailureReason=Invalid Executable. The executable 'Comm.app/Frameworks/hermes.framework/hermes' contains bitcode., NSLocalizedDescription=Validation failed}";
    "iris-code" = "STATE_ERROR.VALIDATION_ERROR";
}
```

I found [this issue](https://github.com/facebook/hermes/issues/1525) on GitHub. It looks like it's resolved in React Native 0.73. @graszka22 is working on upgrading our React Native version, but in the meantime I found a hack in the comments there to unblock the release. I modified it slightly to only run on binaries that we're actually using.