[native] extract constants from `persist.js`
Summary:
Part of [ENG-10477](https://linear.app/comm/issue/ENG-10477/implement-shared-function-to-execute-restore)

To be able to call `restoreBackupData` we need store version. Doing the same approach as for the web for now, but in future this should be defined in lib, but I'll take care of this as part of [Run the migrations during backup restoration](https://linear.app/comm/project/run-the-migrations-during-backup-restoration-52bf6617d3e2/issues) because I am not sure we're ready for this now.

Depends on D14521

Test Plan: Flow

Reviewers: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14522