[lib][native] introduce `fullBackupSupport` flag
Summary:
[ENG-10483](https://linear.app/comm/issue/ENG-10483/introduce-fullbackupsupport-flag).

Unfortunately there is no easy way to deduplicate this. We have `CommConstantsModule`, but we need this to be defined in lib, not only on native. This will be used only by me and @bartek to test backup, so going for the quickest solution for now.

Depends on D14519, D14525

Test Plan: N/A

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek, bartek

Differential Revision: https://phab.comm.dev/D14528